Fix msvc and clang warnings

+++ sse/blake2sp.c
@@ -48,7 +48,7 @@ static int blake2sp_init_leaf( blake2s_state *S, size_t outlen, size_t keylen, u
   P->fanout = PARALLELISM_DEGREE;
   P->depth = 2;
   P->leaf_length = 0;
-  P->node_offset = offset;
+  P->node_offset = (uint32_t)offset;
   P->xof_length = 0;
   P->node_depth = 0;
   P->inner_length = BLAKE2S_OUTBYTES;
